requirements.txt
numpy>=1.26,<2.0
pandas>=2.0
scipy>=1.11
scikit-image>=0.23
scikit-learn>=1.3
tqdm>=4.66
rich>=13.4
PyYAML>=6.0
einops>=0.7
nibabel>=5.2
pydicom>=2.4
SimpleITK>=2.3
timm>=0.9.10
torchio>=0.18.92
monai>=1.3
torchmetrics>=1.0
omegaconf>=2.3.0
opencv-python-headless>=4.8
matplotlib>=3.7
TotalSegmentator==2.9.0
rt-utils==1.2.7
dcm2niix==1.0.20220715
dicom2nifti==2.4.8
python-gdcm==3.0.22
pylibjpeg==1.4.0
ipykernel>=6.24
ruff>=0.9
black>=24.8
pytest>=8.0